SWTASK-302: 하비님 코멘트 반영 - CHANGE _get_ip.py와 _get_os.py 합침 - CHANGE startup_flow.py에서의 get_os 사용 살짝 수정
