+++ release/scripts/startup/abler/startup_flow.py
@@ -35,8 +35,8 @@ from .lib.async_task import AsyncTask
 from .lib.login import is_process_single
 from .lib.read_cookies import *
 from .lib.tracker import tracker
-from .lib.tracker._get_ip import user_ip
-from .lib.tracker._get_os import user_os
+from .lib.tracker._user_info_utils import user_ip
+from .lib.tracker._user_info_utils import user_os
 from .warning_modal import BlockingModalOperator
 import subprocess
 from .lib.version import (
